+++ apps/www/src/registry/default/plate-ui/mention-element.tsx
@@ -3,9 +3,8 @@ import React from 'react';
 import type { TMentionElement } from '@udecode/plate-mention';
 
 import { cn, withRef } from '@udecode/cn';
-import { getHandler } from '@udecode/plate-common';
+import { IS_APPLE, getHandler } from '@udecode/plate-common';
 import { useElement } from '@udecode/plate-common/react';
-import { IS_APPLE } from '@udecode/utils';
 import { useFocused, useSelected } from 'slate-react';
 
 import { useMounted } from '@/registry/default/hooks/use-mounted';