+++ packages/ai/src/lib/transforms/insertAINodes.ts
@@ -6,8 +6,7 @@ import {
   collapseSelection,
   getEndPoint,
   insertNodes,
-  withMerging,
-  withoutMergingHistory,
+  withoutNormalizing,
 } from '@udecode/plate-common';
 
 import { AIPlugin } from '../../react/ai/AIPlugin';
@@ -16,43 +15,23 @@ export const insertAINodes = (
   editor: SlateEditor,
   nodes: TDescendant[],
   {
-    history = 'default',
     target,
   }: {
-    history?: 'default' | 'merge' | 'withoutMerge';
     target?: Path;
   } = {}
 ) => {
   if (!target && !editor.selection?.focus.path) return;
 
-  const insert = () => {
-    const aiNodes = nodes.map((node) => ({
-      ...node,
-      [AIPlugin.key]: true,
-    }));
+  const aiNodes = nodes.map((node) => ({
+    ...node,
+    [AIPlugin.key]: true,
+  }));
 
+  withoutNormalizing(editor, () => {
     insertNodes(editor, aiNodes, {
       at: getEndPoint(editor, target || editor.selection!.focus.path),
       select: true,
     });
     collapseSelection(editor, { edge: 'end' });
-  };
-
-  switch (history) {
-    case 'default': {
-      insert();
-
-      break;
-    }
-    case 'merge': {
-      withMerging(editor, insert);
-
-      break;
-    }
-    case 'withoutMerge': {
-      withoutMergingHistory(editor, insert);
-
-      break;
-    }
-  }
+  });
 };