+++ packages/slate/src/interfaces/history-editor/isHistorySaving.ts
@@ -1,6 +1,6 @@
-import { HistoryEditor } from 'slate-history';
-
 import type { TEditor } from '../editor';
+
+import { HistoryEditor } from '../../slate-history';
 
 /** {@link HistoryEditor.isSaving} */
 export const isHistorySaving = (editor: TEditor) =>