+++ packages/slate/src/interfaces/history-editor/withNewBatch.ts
@@ -1,0 +1,7 @@
+import type { TEditor } from '../editor';
+
+import { HistoryEditor } from '../../slate-history';
+
+/** {@link HistoryEditor.withMerging} */
+export const withNewBatch = (editor: TEditor, fn: () => void) =>
+  HistoryEditor.withNewBatch(editor as any, fn);