+++ packages/slate/src/interfaces/history-editor/withoutMergingHistory.ts
@@ -1,6 +1,6 @@
-import { HistoryEditor } from 'slate-history';
-
 import type { TEditor } from '../editor';
+
+import { HistoryEditor } from '../../slate-history';
 
 /** {@link HistoryEditor.withoutMerging} */
 export const withoutMergingHistory = (editor: TEditor, fn: () => void) =>