+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @udecode/plate-ai
 
+## 39.2.2
+
+### Patch Changes
+
+- [`b6ab2cf4ef9ecad4ccc54b6626d571f8c51c6cff`](https://github.com/udecode/plate/commit/b6ab2cf4ef9ecad4ccc54b6626d571f8c51c6cff) by [@zbeyens](https://github.com/zbeyens) – Fix insert mode
+
+## 39.2.1
+
+### Patch Changes
+
+- [`a17b84f1aa09ac5bcc019823b5d0dfea581ada57`](https://github.com/udecode/plate/commit/a17b84f1aa09ac5bcc019823b5d0dfea581ada57) by [@zbeyens](https://github.com/zbeyens) – Fix lodash imports
+
+## 39.2.0
+
+### Minor Changes
+
+- [#3644](https://github.com/udecode/plate/pull/3644) by [@felixfeng33](https://github.com/felixfeng33) – Add `AIPlugin`, `CopilotPlugin`
+
 ## 39.1.5
 
 ## 39.1.4