+++ packages/slate-react/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @udecode/slate-react
 
+## 39.2.1
+
 ## 39.1.4
 
 ### Patch Changes